+++ firestore-bigquery-export/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Version 0.1.27
 
+fixed - added fix for configuration setup
+
+fixed - update snapshot script
+
+feature - add oldData to the record
+
+fixed - updating table metadata too often
+
 feature - add lifecycle event to export existing documents to Bigquery
 
 ## Version 0.1.26